upstream: b=master,r=c5442ea8137a3320fbc6f8f02698e6693c3d5c12,t=2019-01-18-1732-38672

+++ components/nexus-repository/src/main/java/org/sonatype/nexus/repository/storage/TempBlob.java
@@ -47,6 +47,8 @@ public class TempBlob
 
   private final BlobStore blobStore;
 
+  private boolean deleted = false;
+
   public TempBlob(final Blob blob,
                   final Map<HashAlgorithm, HashCode> hashes,
                   final boolean hashesVerified,
@@ -82,8 +84,12 @@ public class TempBlob
 
   @Override
   public void close() {
+    if (deleted) {
+      return;
+    }
     try {
       blobStore.deleteHard(blob.getId());
+      deleted = true;
     }
     catch (BlobStoreException e) {
       log.debug("Unable to delete blob {} in blob store {}", blob.getId(),
